cleaned up sdk into separate files and with more comments+examples - test_gadgets is now a unittest class - combined listing gadgets and generating a quote for them into single "GadgetsCover" class - made Quotes more OOP for easier subclassing of different modules - Policyholder class as an application of class for type and some validation in python - Resources' call has defaults for sub-classes which can be overridden at call! - main.py has a 'full' process example for gadgets
