+++ tests/context.py
@@ -3,3 +3,4 @@ import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
 from root import insurance
+from root.insurance import InsuranceClient
